feat: Notes Respository created; Migrations generated; pagination middleware added; First notes routing; Jest configured; Types improved

+++ api/src/domain/services/NotesServices.ts
@@ -1,21 +1,21 @@
 import { v4 as uuid } from "uuid"
 
-import { Note } from "domain/entities/Note"
-import { IRepository } from "domain/interface/IRepository"
+import { Note } from "../entities/Note"
+import { INotesRepository, getManyOptions, getManyResponse, updateOneResponse, deleteOneResponse } from "../../infrastructure/shared/interface"
 
 export class NoteServices {
-  private repository: IRepository<Note>
+  private repository: INotesRepository<Note>
 
-  constructor(repository: IRepository<Note>) {
+  constructor(repository: INotesRepository<Note>) {
     this.repository = repository
   }
 
-  async getMany(): Promise<Note[]> {
-    return await this.repository.getMany()
+  async getMany(options: getManyOptions): Promise<getManyResponse<Note>> {
+    return await this.repository.getMany(options)
   }
 
-  async getOne(id: string): Promise<Note> {
-    return await this.repository.getOne(id)
+  async getOne(noteSlug: string): Promise<Note> {
+    return await this.repository.getOneBySlug(noteSlug)
   }
 
   async create(entity: Note): Promise<Note> {
@@ -27,11 +27,24 @@ export class NoteServices {
     return await this.repository.create(parsedEntity)
   }
 
-  async updateOne(id: string, entity: Partial<Note>): Promise<void> {
-    await this.repository.updateOne(id, entity)
+  //TODO: Criar erros personalizdos e os tratamento corretos
+  async updateOne(id: string, entity: Partial<Note>): Promise<updateOneResponse> {
+    const hasEntity = await this.repository.getOneById(id)
+
+    if(!hasEntity) {
+      throw new Error("Not found")
+    }
+
+    return await this.repository.updateOne(id, entity)
   }
 
-  async deleteOne(id: string): Promise<void> {
-    await this.repository.deleteOne(id)
+  async deleteOne(id: string): Promise<deleteOneResponse> {
+    const hasEntity = await this.repository.getOneById(id)
+
+    if(!hasEntity) {
+      throw new Error("Not found")
+    }
+
+    return await this.repository.deleteOne(id)
   }
 }
